Added RO mode support and fixed site partitioning.
Read-only transactions are now supported by multiversion clones on the
sites.

There was an error in how the variables were distributed to the sites
that caused test04 to fail. It is now corrected.

The package is now installable using distutils and also has a script
called repcrec that will run from stdin or local test files.

Captured DEBUG and FULL test outputs for diffs.