Multiversion are local to the TransactionManager.
The Site now uses a token indicating which version is being accessed.
When the token is supplied, the connection semantics of the Site are
ignored and data are always returned when available.

The documentation for Site.multiversion_clone() has a good description
of the broken contracts caused by this behavior. The TransactionManager
is aware of these issues.